requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.1.0
Werkzeug==3.0.1
gunicorn==20.1.0
python-dotenv==1.0.0
